[유치수] [NOBTS] 전체 빌드할 때 중간에 실패가 있으면 중단하도록 함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-server/trunk@2144 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ build.sh
@@ -16,14 +16,60 @@ fi
 
 #server
 mvn clean eclipse:eclipse install package dependency:copy-dependencies -Dmaven.test.skip $PROFILE
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
 
 rm -fr $DEPLOY_DIR
 mkdir -p $DEPLOY_DIR/lib
 
 cp ./target/pinpoint-collector-$VERSION.jar $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./target/dependency/*.jar $DEPLOY_DIR/lib
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./runscript/* $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./target/classes/applicationContext*.xml $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./target/classes/hbase.properties $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./target/classes/log4j.xml $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
+
 cp ./target/classes/pinpoint-collector.properties $DEPLOY_DIR
+rc=$?
+if [[ $rc != 0 ]] ; then
+        echo "BUILD FAILED $rc"
+        exit $rc
+fi
